Switch to a new OBS websocket implementation, not at parity yet
